requirements.txt
psycopg2-binary
flask
Flask-SQLAlchemy
flask-migrate
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-jaeger